% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbDecoda.R
\name{mbDecoda}
\alias{mbDecoda}
\title{a debiased approach to compositional data analysis specifically designed for microbiome surveys.}
\usage{
mbDecoda(
  count,
  x,
  Gamma = NULL,
  W = NULL,
  interval.prob = 0.5,
  signif = 0.05,
  d_init = F,
  prev.cut = 0.05,
  maxit = 100,
  reltol = 1e-05,
  adjust = "BH"
)
}
\arguments{
\item{count}{matrix representing observed OTU table. Row: taxa; column: samples. NAs are not expected in OTU tables}

\item{x}{matrix representing the variable of interest}

\item{Gamma}{matrix representing the other covariates of NB model to be adjusted. Default is NULL}

\item{W}{matrix representing the variable of the zero-inflated model. Default is NULL}

\item{interval.prob}{interval size of MCI. Default is 0.5}

\item{signif}{the level of significance to count the reject number, default is 0.05}

\item{d_init}{the initial value of d, default is F}

\item{prev.cut}{a numerical fraction between 0 and 1. Taxa with proportion of zeroes less than prev.cut will be excluded in the analysis. Default is 0.05}

\item{maxit}{the maximum number of iterations for the E-M algorithm. Default is 100}

\item{reltol}{Default is 1e-5}

\item{adjust}{method to adjust p-values by. Default is "BH". Options include "BH", "hochberg", "hommel", "bonferroni", "holm", "BY", "fdr", "none".}
}
\value{
a list with three components:
\itemize{
\item ZINB
\itemize{
\item alpha, parameter estimates of the zero-inflated model
\item alpha_p, p-values of the zero-inflated model parameters
\item phi, dispersion parameter estimates of the NB (Negative Binomial) distribution
}
\item Confounders
\itemize{
\item beta, estimates of parameters of Gamma in the NB model
\item beta_p, p-values of beta
\item beta_sd, standard error of beta
}
\item Bias_correct
\itemize{
\item delta_bias, estimates of delta before MCI
\item bias, estimate of bias using MCI
\item sd_bias, standard error of bias
\item DAA, differential abundance analysis result on absolute level
}
}
}
\description{
a debiased approach to compositional data analysis specifically designed for microbiome surveys.
}
\examples{
data = ZINB_sim(seed=1,K=50,n1=20,n2=20,p=0.1,bias ="small",zi=0.3,confounder =F)
group = data[["grp"]]
count = data[["count"]]
out = mbDecoda(count, x=group)
}
