% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINB_sim.R
\name{ZINB_sim}
\alias{ZINB_sim}
\title{Function to simulate ZINB data}
\usage{
ZINB_sim(
  seed,
  K,
  n1,
  n2,
  p,
  bias = c("large", "small"),
  zi = 0.3,
  overdisp = 3,
  confounder = c("TRUE", "FALSE")
)
}
\arguments{
\item{K}{numbers of taxa}

\item{n1}{sample size of group1}

\item{n2}{sample size of group2}

\item{p}{the propotion of the differential taxa}

\item{bias}{bias size, "small" or "large"}

\item{zi}{the propotion of zero inflation, default is 0.3}

\item{overdisp}{default is 3}

\item{confounder}{"TRUE" or "FALSE"}
}
\description{
Function to simulate ZINB data
}
\examples{
ZINB_sim(seed=1,K=50,n1=20,n2=20,p=0.1,bias ="small",zi=0.3,confounder =F)
}
